move some profile APIs into profile controller
